RDKB-37175 : WebConfig: Mgmt Frame Pwr not configured properly
Reason for change:  Definition of parameter in hal is unsigned int
and hence signed values are failing in driver
Test Procedure: Attached to the ticket
Risks: Low

(cherry picked from commit 3b52a576ca6817ee4109dc75d8ed6504c8839910)

+++ source/TR-181/ml/cosa_wifi_dml.c
@@ -8650,7 +8650,7 @@ AccessPoint_SetParamIntValue
             CcspWifiTrace(("RDK_LOG_ERROR, %s Unable to get VAP info for apIndex:%lu\n", __FUNCTION__, apIndex));
             return FALSE;
         }
-        if ( vapInfo->u.bss_info.mgmtPowerControl == (UINT)iValue )
+        if ( vapInfo->u.bss_info.mgmtPowerControl == iValue )
         {
             return  TRUE;
         }
